Refactoring, split up utils package
