+++ src/cli/index.ts
@@ -1,0 +1,2 @@
+export { createMidiStatusReport } from "./createMidiStatusReport";
+export { reportMidiStatus } from "./reportMidiStatus";
