+++ src/midi/index.ts
@@ -1,0 +1,6 @@
+export { createMidiMessageRouter } from "./createMidiMessageRouter";
+export { getMidiChannel } from "./getMidiChannel";
+export { getPortIndex } from "./getPortIndex";
+export { initPort } from "./initPort";
+export { isControlChange } from "./isControlChange";
+export { isSketchSwitch } from "./isSketchSwitch";
