+++ src/utils/index.ts
@@ -1,10 +1,2 @@
-export { createMidiStatusReport } from "./createMidiStatusReport";
-export { createMidiMessageRouter } from "./createMidiMessageRouter";
 export { formatHex } from "./formatHex";
-export { getMidiChannel } from "./getMidiChannel";
-export { getPortIndex } from "./getPortIndex";
-export { initPort } from "./initPort";
-export { isControlChange } from "./isControlChange";
-export { isSketchSwitch } from "./isSketchSwitch";
-export { reportMidiStatus } from "./reportMidiStatus";
 export { withFatalErrorOnNull } from "./withFatalErrorOnNull";
